remove all zombie code
only the admin component is needed

test plan: browse to the the cred explorer and the admin app and ensure
both are visible
