+++ src/ui/Main.js
@@ -3,17 +3,14 @@
 import React from "react";
 import {Route, Switch} from "react-router-dom";
 import ExplorerApp from "./components/ExplorerApp";
-import InitiativesEditor from "./components/InitiativesEditor";
+import AdminApp from "./components/AdminApp";
 
 export default class Main extends React.PureComponent<{||}, {||}> {
   render() {
     return (
       <Switch>
         <Route exact path="/" render={(props) => <ExplorerApp {...props} />} />
-        <Route
-          path="/editor"
-          render={(props) => <InitiativesEditor {...props} />}
-        />
+        <Route path="/editor" render={(props) => <AdminApp {...props} />} />
       </Switch>
     );
   }
